Short-circuit consensus
Large file values cause the consensus code to fail, due to overflowing
udp packets. If a packet like this is encountered all write ops come
to a halt and doozerd is not fully operational anymore. As we run
doozerd in a single machine setup we're not dependent on the
consensus code.

This change short-circuit that by passing the same channel as in
and out channel to the proposer used a by the peer. Additionally we
get rid of the marshaling of the packets into protobuf as it never
goes on the wire.